+++ src/cppwasm/TestHelpers.hpp
@@ -7,14 +7,14 @@
 
 using namespace std;
 
-inline void insertSin(vector<complex<double>> &buff, int fps, double maxAmplitude, double freq)
+inline void insertSin(vector<complex<double>> &buff, size_t fps, double maxAmplitude, double freq)
 {
     auto sz = buff.size();
-    for (int x = 0; x < sz; x++)
+    for (size_t x = 0; x < sz; x++)
         buff[x] += maxAmplitude * sin(M_PI * x * 2 * freq / fps);
 }
 
-inline vector<complex<double>> generateSin(int size, int freq)
+inline vector<complex<double>> generateSin(size_t size, int freq)
 {
     vector<complex<double>> toGen(size, 0);
     insertSin(toGen, size, 10, freq);
